apps/api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
strawberry-graphql[fastapi]>=0.213.0
supabase>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Database & ORM
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# Development dependencies
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0
